adding inline comments + edits here and there
