+++ packages/redeem-failed-retryable/.env-sample
@@ -9,7 +9,3 @@ L2RPC="https://rinkeby.arbitrum.io/rpc"
 
 # Ethereum RPC i.e., for rinkeby https://rinkeby.infura.io/v3/<your infura key>
 L1RPC=""
-
-
-# Address of the Arbitrum Inbox on the L1 chain
-INBOX_ADDR=0x578bade599406a8fe3d24fd7f7211c0911f5b29e
