+++ packages/redeem-failed-retryable/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "failed-retryables",
+  "name": "redeem-failed-retryable",
   "license": "Apache-2.0",
   "version": "1.0.0",
   "scripts": {
